Removing warnings from compile.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@506 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/Models/MultiBodyModel.cpp
@@ -165,7 +165,7 @@ MultiBodyModel::AddJoint(ConnectionModel* _c, int _indexBase, int _indexJoint, i
 
 void
 MultiBodyModel::AddBody(BodyModel* _b, int _index){
-  if(_index<=m_bodies.size())
+  if(_index <= (int)m_bodies.size())
     m_bodies.insert(m_bodies.begin()+_index, _b);
   else
     m_bodies.push_back(_b);
